MB-28453: memcached_testapp: Add Snappy variant of XATTR tests
Add a new paramter to TestappClient tests on if the client supports
Snappy or not.

Make use of this new paramtere in the XATTR tests to run with both
Snappy enabled and disabled.

To increase the amount of test coverage we can acheive via testapp,
change to use SET_WITH_META to store test documents which have
XATTRs. By using SET_WITH_META we can compress the complete
value+XATTR body; otherwise when the testapp client uses the
individual sub-doc operations to store the test document we can only
store uncompressed documents.

Change-Id: I7b177a0606b24f09ba9060ada91d67f1528e6096
Reviewed-on: http://review.couchbase.org/90446
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: David Haikney <david.haikney@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>