MB-28492: Allow running verbose mode from testapp tests
Due to the layered architecture of spdlogger where we build
up a chain of sinks where we send the log message, each of
these sinks may filter only a subset of the messages.

When running in production we don't want the console logger
to log below the warning level, as ns_server is the consumer
and it can't keep up with the amount of data we would otherwise
log. ns_server's solution to that is to "drop" the log, so
we'd rather just log at a rate it can consume.

Running in unit-test-mode we do however want to be able to
send more output to the console so we need to set the filter
level to trace (note that the filter set by the logger
on top would cause the events to be filtered out unless
verbosity is bumped)

Change-Id: I71e4bfce2fbb258cad9432a0ec90c012c39524bf
Reviewed-on: http://review.couchbase.org/90496
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>