default_engine: Use zero-length / RAW_BYTES for deleted docs
When allocating an item to represent a deleted document, use
datatype:RAW_BYTES and a length of zero for the empty body. This is
necessary as if the previous body was Snappy-compressed and we
preserve the datatype; we end up with a deleted document marked as
snappy, but which we cannot decompress. As a result operations like
XATTR on deleted documents fail.

Change-Id: I336d774442b06b58c7fae2b2742f6f1e6fc7bbe3
Reviewed-on: http://review.couchbase.org/90523
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>