MB-28289: Allow GetFailoverLog on a regular-MCBP connection
Currently GetFailoverLog is a DCP command, thus a client needs to open
a DCP connection to get the failover log for a given VBucket.
Also, the DCP GetFailoverLog command requires the client to open a
DcpProducer on the node; the connection is closed otherwise.
We introduce a new MCBP GetFailoverLog command so that a client can
retrieve the Failover Log for a given VBucket via a regular MBCP
(non-DCP) connection. We keep the old DCP command for backward
compatibility.

Change-Id: I21445149f509636dbc20059e1c8d3f36af8b38c0
Reviewed-on: http://review.couchbase.org/90527
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>