MB-28524: Use uncompressed value when merging existing XATTRs
When an existing document with XATTRS is replaced with a new value, we
merge the XATTRs into the new value - see
getExistingItemToPreserveXattr().

However, if the new value was Snappy-compresed by the client, we fail
to decompress it before merging with the XATTRs. The effect of this is
the document is essentially corrupted; it's marked as
datatype=Snappy|XATTR, but only the value is Snappy-compressed, not
the whole XATTR+Value. As such it cannot be correctly read.

Address by using the decompressed form of such input documents when
creating the combined value+XATTR.

Change-Id: I1b5811e1346c2d69bf1a1905605434706032c14f
Reviewed-on: http://review.couchbase.org/90541
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>