MB-28524: Decompress existing XATTR when merging with new value
When an existing document with XATTRS is replaced with a new value, we
merge the XATTRs into the new value - see
getExistingItemToPreserveXattr().

However, if the existing value (including XATTRs) Snappy-compresed, we
fail to decompress the XATTRs before merging with the new value.

The effect of this is the document is corrupted; it's marked as
datatype=XATTR, but the XATTR part is actually snappy compressd.
As such it cannot be correctly read.

Address by using the decompressed form of the existing XATTRs when
creating the combined value+XATTR.

Note regression test follows in subsequent patch.

Change-Id: If5efcc3eff32c3c8aa828899e1da5b656bb85404
Reviewed-on: http://review.couchbase.org/90598
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>