Add logging for socket errors during create and connect
And if MEMCACHED_UNIT_TEST is set in enviroment we'll
retry connection attempts to the server if connect
returns EADDRINUSE. We've seen that problem causing
unit tests to fail on our Windows CI system. My
understanding of the problem is that we're "running out of
sockets (random ports)" due to the fact of all of these
shortlived connections all of our unit tests make (every
test case causes at least 4 new connections to be established
and we've got a fair amount of test cases running). The
kernel will keep them around for some time after they close
but before we may reuse them.

Change-Id: Ib3ea6502b3ce861963028cfd7298c52b13844624
Reviewed-on: http://review.couchbase.org/90601
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>