MB-28651: Ensure audit.log exists as much as possible
The changes documented below are to ensure that the audit.log file
almost always exists and open available for writing.

Note: It is not possible for the file to always exist because when
we rotate the file we close the current one, rename and open the new
one.

The changes are as follows:

Create the audit.log file immediately after a file rotation occurs (as
opposed to when a new audit event is recieved).

If the log file is due to rotate, but it is currently empty,
instead of closing the file, keep it open and update the open_time
so that the next rotation will occur at the correct time.

If there is a failure writing to disk, which causes us to close (and
possibly rotate) the audit file, make sure that the file exists and is
re-opened immediately afterwards.

Change-Id: Ib93abf1d45eb36c15b6f2dee438f9956894fff58
Reviewed-on: http://review.couchbase.org/90599
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>