Terminate the tests if the server is gone
If the server dies while running the tests we can might as well
terminate the test application.

We used the ASSERT_* macros in the code to check for this,
but gtest doesn't bypass the test if one of these fails
during SetUp. This cause a LOT of output on the console as
every test case will fail from this point on).

Change-Id: I0b290703934ea1cdb1a0dc6cd17410b513e17476
Reviewed-on: http://review.couchbase.org/90652
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>