MB-28573: The consumer must ack even if an error occurred
Even if a mutation hit the error where an allocation failed or the
ExtendedMetaData was invalid, it still needs to ack back to the
producer, otherwise we risk indefinitely pausing replication.

To address this issue, introduce a RAII helper class in the
DcpConsumer which will perform the flowControl increment when
destructing (providing release() wasn't called). Now DcpConsumer
functions can just create the UpdateFlowControl object and allow RAII
to always update the flow control (and notify the consumer) when an
error occurred or successful.

Note the ufc.release() cases are because when TMPFAIL is received it
means the dcp message has been buffered for future processing, the
future processing will handle the flowControl update, so we must call
release and avoid a double count.

Change-Id: I32aff3c472b9a8d794815da83a1cdef39ecb32f4
Reviewed-on: http://review.couchbase.org/90654
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>