ValueTest: Use factory to create object instances
The initialFreqCounterForTemp() test is typed on two types -
StoredValue and OrderedStoredValue. Therefore one must use the factory
method appropriate to each type to create additional objects. If not,
then we end up trying to refer to a StoredValue object via an
OrderedStoredValue pointer. This triggers an essertion in debug
builds when trying to delete the OrderedStoredValue:

    Assertion failed: (!hook.is_linked()), function destructor_impl, ...

Change-Id: I8a2875c993b32e1417c4bd19cf4e8c340320c530
Reviewed-on: http://review.couchbase.org/90674
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>