ValueTest: Don't attempt to copy StoredValue objects
StoredValue objects do not support copying directly from other
StoredValue objects - only OrderedStoredValue is supported (as
Ephemeral requires that to deal with adding Values when a range read
is in progress).

As such, modify the copyStoreValue test to only instantiate for
OrderedStoredValue - and use the factory to copy correctly.

Note that MockStoredValue is now dead with this change (it didn't
actually do anything previously anyway), so remove it.

Change-Id: Id213fac8bebe0118045f0f3fe29d1cefb3fb32b4
Reviewed-on: http://review.couchbase.org/90675
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>