MB-27197 CouchRocks: Update VBState without reading high seqno from disk
When we commit a batch of mutations, we do not need to read the highest
persisted seqno from disk to check that it is correct, as that is ensured
by rocksdb::DB::Write() executing successfully.
So, we can just update the VBucket state with the highest seqno that we
find in the commit batch.

Change-Id: I1b443d389232144a98c8b9c2772f92e370df9164
Reviewed-on: http://review.couchbase.org/90786
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>