MB-28629: Improve logging when failing to set up server socket
Previously we did not print out which port we failed to configure
during startup. Instead the log looked like:

    INFO Enable management port(s)
    INFO Enable user port(s)
    CRITICAL Failed to create listening socket

This patch extends this to:

    INFO Enable management port(s)
    INFO Enable user port(s)
    CRITICAL Failed to create IPv4 port for "*":11210
    CRITICAL Failed to create IPv6 port for "*":11210
    CRITICAL Failed to create listening socket

This gives deeper insight in why we failed to start. It is
still considered a fatal error if we fail to create all ports
specified by ns_server as we don't really have a way to tell
ns_server which port we failed to set up (and it can't tell
us to try again).

Change-Id: I35dbdc23f4dae08a9f196b113d11bd1b8e184c9a
Reviewed-on: http://review.couchbase.org/90791
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>