MB-28633: Don't create a temp std::string in ScanContext::manageSeparator
When creating a DocKey (which is just a view on the underlying data)
in ScanContext::manageSeparator(), don't create a temporary
std::string. This is illegal as when the string goes out of scope
(after calling DocKey::make() at line 34) the resulting 'cKey2' will
be pointing at now-freed heap (assuming no small-string optimization), 
or invalid stack (with small-string optimization).

Change-Id: I055d0d52277974b25b328f5fc3ca96b6fba9b6fa
Reviewed-on: http://review.couchbase.org/90796
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>