MB-28630: Try to catch std::bad_function_call during privilege check
The std::bad_function_call is thrown when a std::function is
called but no function is specified (whithout any sensible
context for us to locate where it comes from). My gut feeling
tells me this is because the system is in a hosed state (most
likely cookie corruption), but in order to try to locate where
it is being thrown we need to start narrowing the places to
look.

Change-Id: I8fdbab0f0baf0eeecf38e910b70153624de15ef2
Reviewed-on: http://review.couchbase.org/90844
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>