MB-28630: Extend log info for exception in runloop
If an exception is thrown while executing stuff for a
connection it is being caught and the connection is being
shut down.

Log the entire cookie information if the exception occurs
during packet execution. For the other states include
the state name in the exception.

In addition to that catch std::bad_alloc and ignore
logging if it is being thrown while adding log entries
(to avoid crashing the server)

Change-Id: Ib7700e5b7099ccaa054dfd2f61f2784f76399815
Reviewed-on: http://review.couchbase.org/90843
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>