Explicitly enable ASan for KV-Engine targets needed for cluster_run
To be used with the new CB_ADDRESSSANITIZER=2 mode; enable ASan for
the targets needed to run KV-Engine as part of a larger cluster-run;
but without having to enable for all targets.

For example; enabling ASan for libcouchstore.so causes problems for
the view engine as the main executable which loads libcouchstore
(Erlang VM) doesn't have ASan enabled and hence it crashes on startup.

Change-Id: Ieb77f0622d581e5d7767b2a4fba66853416dd342
Reviewed-on: http://review.couchbase.org/90866
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>