MB-28457: Revert max_checkpoints to 2
Investigation on the replication slowdown from this MB indicates that
it is somehow related to increasing the maximum number of checkpoints
from 2 to 100.

Analysis continues as to exactly why this causes the regression, but
to restore the performance and avoid issues in other test, revert back
to max of 2 checkpoints for now.

Change-Id: I1eefdce8dd3a58c3bd939ff71b42a0faaff97855
Reviewed-on: http://review.couchbase.org/90864
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>