MB-28670: UBSan: CouchKVStore: Change dbFileRevMap to shared_ptr
UBSan reports the following undefined behaviour when constructing a CouchKVStore object:

    couch-kvstore.cc:305:20: runtime error: member access within address 0x000107b17000 which does not point to an object of type 'CouchKVStore'
    0x000107b17000: note: object has invalid vptr
     00 00 00 00  00 70 b1 07 01 00 00 00  00 70 b1 07 01 00 00 00  00 00 00 00 00 00 00 00  18 70 b1 07
                  ^~~~~~~~~~~~~~~~~~~~~~~
                  invalid vptr

The issue is that we are passing a reference to a class member
variable (fileRevMap) before it has been constructed.

The reason for the code to do this was to handle Read-Only and
Read-Write CouchKVStore instances; which share a single revision map
object between them - the r/w instance uses it's own (concrete)
revisionMap member variable; whereas the r/o uses a reference to the
R/W's member.

To avoid the undefined behaviour, restructure so we use a shared_ptr
to jointly own the revision map; and create the map explicitly in the
constructor (before we delegate to other constructors).

Change-Id: I6830d92912bb866ad6bc7bdd59ae6996d8b4950f
Reviewed-on: http://review.couchbase.org/90926
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>