MB-28671: UBSan: Remove static_cast<EPBucket&> before object constructed.
UBSan reports the following undefined behaviour when constructing a KVShard object:

    kvshard.cc:64:20: runtime error: downcast of address 0x000107bb2000 which does not point to an object of type 'EPBucket'
    0x000107bb2000: note: object is of type 'KVBucket'
    #0 0x10080f058 in KVShard::KVShard(unsigned short, KVBucket&) kvshard.cc:64

KVShard's constructor takes a KVBucket& as parameter; which is casted
to an EPBucket& if the config indicates persistence is enabled. While
the object /is/ indeed of type EPBucket, KVShard's constructor is
called indirectly from EPBucket's constructor; and as such the
EPBucket object hasn't been fully constructed yet.

To avoid this problem, remove the KVBucket& parameter, and instead
have a seperare, explicit enablePersistence() method which performs
the work which was previously done in the constructor.

Change-Id: Iff1b2bd38c3af257fc51f97de4050da85f3b3ac9
Reviewed-on: http://review.couchbase.org/90927
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>