MB-28746: UBSan: Ignore incorrect static_cast to derived
As reported by UBSan:

    dcp_test.cc:2079:13: runtime error: downcast of address 0x000107b95000 which does not point to an object of type 'MockDcpConnMap'
    0x000107b95000: note: object is of type 'DcpConnMap'

This cast is undefined behaviour - the DCP connection map object is of
type DcpConnMap; so it's undefined to cast to MockDcpConnMap. However,
in this instance MockDcpConnMap has identical member variables to
DcpConnMap - the mock just exposes normally private data - and so this
/seems/ ok.

Fixing it correctly would involve invasive changes to
EventuallyPersistentEngine to expose the DcpConnmap, therefore allow
it in general, but skip this test under UBSan.

Change-Id: I75afbd586579dd79e6fc9818a4a90d515f3e9228
Reviewed-on: http://review.couchbase.org/91020
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>