MB-28744: UBSan: Remove static_cast from BinprotCommandT
UBSan complains about incorrect cast to derived in BinprotCommandT:

    client_mcbp_commands.h:217:16: runtime error: downcast of address 0x7ffeefbfe5d0 which does not point to an object of type 'BinprotSaslAuthCommand'
    0x7ffeefbfe5d0: note: object is of type 'BinprotCommandT<BinprotSaslAuthCommand, (unsigned char)33>'

I'm not totally sure this warning is accurate; the use of
static_cast<> here (CRTP) should be valid. However it (a) possibly
isn't, or (b) we are missing some RTTI information to allow UBSan to
check the cast.

Either way, we don't really make use of the static polymorthism here -
it was added to support a fluent interface for BinprotCommand, but we
don't really use that very much. Therefore remove these methods (and
the use of static_cast) to avoid the warning.

Change-Id: I89b3d32b42019f30e8091660dc8b70f48597a57c
Reviewed-on: http://review.couchbase.org/91022
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>