MB-28493 CouchRocks: Fix 'pure virtual method called' in ep_testsuite
In RocksDBKVStore we declare the following member variables in the
following order:

- std::unique_ptr<rocksdb::DB> rdb;
- rocksdb::ColumnFamilyOptions defaultCFOptions;
- rocksdb::ColumnFamilyOptions seqnoCFOptions;
- std::shared_ptr<rocksdb::Cache> blockCache.

The member declaration order leads to the inverse destruction sequence
'blockCache->seqnoCFOptions->defaultCFOptions->rdb'.
At first look, the issue seems the following:
- rdb (in some internal structure) keeps a rocksdb::Cache& to
  *blockCache (note that a Ref& does not incrememnt the ref-count of
  a shared_ptr)
- given the destruction order, blockCache is releases (ref-count=0)
- rdb has still rocksdb::Cache&, which is now referring to released
  memory
When a method of rocksdb::Cache& is called, then we have the 
'pure virtual' error.

A couple of open RocksDB issues have been identified as possible cause:
- https://github.com/facebook/rocksdb/issues/3609
- https://github.com/facebook/rocksdb/issues/3534

As a temporary workaround we destroy 'rdb' as first step in the
destruction sequence.

Change-Id: I1abd52ffc55c3d8ac41e072b3097541df1d37532
Reviewed-on: http://review.couchbase.org/91010
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>