MB-28327: Modify mapping from frequency count to 2-bit LRU
When converting from the frequency counter to the original 2-bit_lru
we want the intial frequency to map to the 2nd coldest of the 4
available LRU states.  As this means we will differentate between
items transferred over DCP from memory as opposed from disk.
Therefore the mappings from frequency counter to 2-bit_lru have been
changed such that:

0-3 => 3 /* coldest */
4-31 => 2
32-63 => 1
64->255 => 0 /* hottest */

In addition the initial frequency count is decreased from 5 to 4 so that
it is the first value that maps to the 2nd coldest of the 4 available
LRU states.

Change-Id: I28fe72131fb8644e8ad1d783f21387e36e439854
Reviewed-on: http://review.couchbase.org/91013
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>