MB-28741: UBSan: fix invalid reads of unscoped enums
UBSan reports an error when we (deliberately) encode an invalid value
for in an enum - for example protocol_binary_subdoc_flag in test code:

    kv_engine/protocol/connection/client_mcbp_commands.cc:510:36:
    runtime error: load of value 255, which is not a valid value for type 'protocol_binary_subdoc_flag'

This is because we use an unscoped enum (protocol_binary_subdoc_flag)
without an underlying fixed type, and we deliberately assign a value
which is outside the set of defined subdoc_flag values.

As per the C++ standard[1] (and highlighted by INT50-CPP[2]), this
results in undefined behaviour, due to how the valid values of the
enum are defined:

    10.2.7: For an enumeration whose underlying type is fixed, the
    values of the enumeration are the values of the underlying
    type.
    Otherwise, for an enumeration where emin is the smallest
    enumerator and emax is the largest, the values of the enumeration
    are the values in the range bmin to bmax, defined as follows: Let
    K be 1 for a two’s complement representation and 0 for a one’s
    complement or sign-magnitude representation. bmax is the smallest
    value greater than or equal to max(|emin| − K, |emax|) and equal
    to 2M − 1, where M is a non-negative integer. bmin is zero if emin
    is non-negative and −(bmax + K) otherwise. The size of the
    smallest bit-field large enough to hold all the values of the
    enumeration type is max(M, 1) if bmin is zero and M + 1
    otherwise. It is possible to define an enumeration that has values
    not defined by any of its enumerators. If the enumerator-list is
    empty, the values of the enumeration are as if the enumeration had
    a single enumerator with value 0.

(The above basically says that for an unscoped enum without an
underlying fixed type, the range of valid values are derived from the
set of enumerations in the type).

Additionally, when converting to an enum:

    8.2.9.10: A value of integral or enumeration type can be
    explicitly converted to an enumeration type. The value is
    unchanged if the original value is within the range of the
    enumeration values (7.2). Otherwise, the resulting value is
    undefined.

To summarise - for an unscoped enum without a fixed type, it is
undefined to assign a value which is outside the range of the defined
enumerations.

The same isn't the case for scoped enums, or for unscoped enums with a
fixed type. While changing to scoped enums is generally a Good Idea,
that's a more invasive change at this point, so instead we choose to
stick with an unscoped enum, specify a fixed underlying type.

[1]: http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/n4659.pdf
[2]: https://wiki.sei.cmu.edu/confluence/display/cplusplus/INT50-CPP.+Do+not+cast+to+an+out-of-range+enumeration+value

Change-Id: Ic14fc597ef612fcbf9f5dbc220497571c4a5b103
Reviewed-on: http://review.couchbase.org/91096
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>