Cleanup: make TestappTest::getConnection() predictable
getConnection() used to return whatever was stored in "front" of
the connection map, which could be PLAIN IPv4/6 or SSL IPv4/6.
This caused some tests to fail (cert test when trying to set
certificate was mandatory) and the front of the map contained
an ssl connection.

The TestappTest is not parameterized with ipv4/6 and plain/ssl
so it should always just use a plain IPv4 connection (and let
people require a connection of the required type if they need
something else)

Change-Id: I2c5f294e09198273973529354bae5d44d7c4218f
Reviewed-on: http://review.couchbase.org/91144
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>