MB-28778: Increase #worker threads to 7/8 of CPUs
The number of frontend threads is based on the number of CPU
cores. This is currently 3/4 of the number of CPU cores for systems
with at least 5 cores).

However, this was set when systems had smaller CPU counts. Systems
with high core counts are more and more common, and limiting the
front-end thread count to 75% of the cores leaves a signficnant amount
of resource unused for larger systems.

Increase to a higher percentage on larger systems - create #CPUs * 7/8
(87.5%) for systems with 5 or more cores.

Example of change in worker threads for some common CPU counts:

  CPUs   Spock (3/4)   Vulcan (7/8)
  ---------------------------------
     4             4              3
     8             6              7
    16            12             14
    24            18             21
    32            24             28
    48            36             42
    64            48             56

Change-Id: I23d967a96bde7da819508a429700fa8dcccdef41
Reviewed-on: http://review.couchbase.org/91182
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>