MB-28206: Stop 'get_meta' from affecting 'get' stats
When getMeta was moved into memcached, we introduced behaviour which
meant that these operations were being included in the regular get stats
however, this caused these operations to be counted twice in ns_server
and so the data shown to the user was higher then reported (seen when
using XDCR). As this behaviour is not currently replicated by setMeta,
stop this from happening to A) maintain consistency and B) stop the
incorrect reporting of the get operations count.

This change also adjusts the existing unit test from checking that
the stats were incremented, to ensuring that they remain at 0
throughout the getMeta operations.

Change-Id: I2fcab2ae2273bb2907a908282435869b12ba1395
Reviewed-on: http://review.couchbase.org/91413
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>