Create eligibleForPageOut function
The hifi_mfu eviction algorithm needs to identify whether a
StoredValue is eligible to be pagedOut of memory without actually
paging it out.

Previously we were using the StoredValue eligibleForEviction for
eviction function however this was found not to be valid for ephemerial
buckets, as the isDirty state is always true.

Therefore we factor out the logic responsible for determining if a
StoredValue is eligible to be paged out from both the couchbase
(persistent) and ephemerial versions of the pageOut function and put it
into a new function called eligibleForPageOut.

The hifi_nfu eviction algorithm is changed to use the current vbucket's
eligibleForPageOut function.

Change-Id: Ib4dd4cfc984a93408a4f9cc41ccf3eb391a470d8
Reviewed-on: http://review.couchbase.org/91512
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>