MB-25562 [1/2] Include the body CRC32C in the '$document' Virtual XAttr
Sync Gateway uses the CAS for checking that it has a perticular version
of a given document. Given the way XATTR is implemented, if either the
document body or only a XAttr changes then the document CAS changes. If
another client (e.g. Eventing) amends a XAttr concurrently, then the
Sync Gateway may have to recalculate the mobile meta.
So, using the CAS for synching is inefficient when multiple actors
compete to the last entity to update a document.

With this change we expose the body CRC32C checksum via the new
'$document.body_crc32c' path. XAttr clients can use the body CRC32C to
check if the body of the document (XAttrs excluded) has changed.

In a follow-up patch we also expose a new Xattr macro to allow clients
to stamp the body CRC32C into a client-provided XAttr.

Change-Id: Iab6aa6982434b8b975c0634b30d39e4a22eb3144
Reviewed-on: http://review.couchbase.org/91594
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>