MB-28387: Increase flusher_batch_split_trigger to 1M
The flusher_batch_split_trigger was initially added to address issues
with very large flush batches (16M+ items) consuming large amounts of
memory and triggering OOM (see MB-26021 / MB-26833). It was intially
defaulted to 10,000 - i.e. we should split a flush (where valid) if
the number of items to be flushed exceeded 10,000.

However, this has caused a number of performance tests to regress -
see MB-28387. The cause of the regression is that by limiting flushes
to ~10,000 items, the overall throughput of couchstore / flusher is
contrained; as there is essentially a high per-commit cost in
couchstore.

To address this, the flusher_batch_split_trigger is increase to
1,000,000. Testing so far shows this is an order of magnitude higher
than needed to address the regression on the tests so far (100k is
sufficient), but we don't want to overly constrain larger systems /
workloads.

Change-Id: Ibdfc536a859865eeb494ec2f9e1e521c1ebf4ef0
Reviewed-on: http://review.couchbase.org/91679
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>