MB-28940: Add benchmark for StatisticalCounter::generateValue
It is suspected that the generateValue function in StatisticalCounter
is responsible for a reduction in performance.  In particular
contention on a lock used within the function is expected to be the
cause.

The patch contains a benchmark for evaluating how long the generateValue
takes when running in a multi-threaded environment.

Example output from the benchmark is provided below.

./ep_engine_benchmarks --benchmark_filter=BM_SaturateCounter
Run on (8 X 2500 MHz CPU s)
2018-03-28 17:04:15
--------------------------------------------------------------------
Benchmark                             Time           CPU Iterations
--------------------------------------------------------------------
BM_SaturateCounter/threads:1         28 ns         28 ns   22953679
BM_SaturateCounter/threads:8       3723 ns       3853 ns     180248

Change-Id: If92f2f002006daa987d7e480a3de59d26a66247f
Reviewed-on: http://review.couchbase.org/91749
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>