MB-25562: Fix the '${Mutation.value_crc32c}' macro padding
Before expanding a macro, we generate a padding and we write it in the
document body. The size of the padding must be equal to the size of the
final macro expanded value.

The 'generate_macro_padding()' function generated always a 8-byte
padding. That was correct for CAS and SEQNO macros (as they expand to
8-byte value)s, but that is wrong for CRC32C 4-byte values.
E.g., expanding a CRC32C checksum into a 8-byte padding results in
something like '\"0xbcd1a5ab"4aa1e75\"'.

To fix, we ensure that the padding is always the same size of the final
expanded value.

Change-Id: Ic5d2f5b40d05eaf3f1b6f5bea0f677eae1cdb5c9
Reviewed-on: http://review.couchbase.org/91825
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>