Remove default parameter values from VBCBAdaptor
The VBCBAdaptor constructor take multiple parameters with the last two
(sleep and shutdown) being given default parameters.  The sleep
parameter is always provided and therefore there is no point in
providing a default value.

The VBCBAdaptor constructor is only invoked from two places.  In one
case the shutdown parameter is provided (true).  In the other case the
default value (false) is used.  Instead of relying on the default
value we can just pass in false.  This is useful for a future patch that
extends the VBCBAdaptor constructor parameters.

Change-Id: I1b5611a81ffdacc0464b67cff08571e279f2b68f
Reviewed-on: http://review.couchbase.org/92023
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>