Add getNumOfVBucketsInState method
Adds functionality to KV buckets to be able to retrieve the number
of underlying vbuckets that are in a given state, i.e. active, replica,
dead, and pending.

This functionality will be used in the future by the hifi_mfu hash
table eviction algorithm to determine the percentage of replica
documents to evict.

Change-Id: Id19e3d28ca3c99306dddc12b7dfe8ee50d84926c
Reviewed-on: http://review.couchbase.org/92042
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>