MB-29108: Create connection stats task to fix concurrent access conditions
Creates a new StatsTask class, which is to be used from the Stats Command
Context.

Implement a StatsTask_ConnectionStats class which is used to get the
connection stats, and use this in place of connection_stats function
in connections.cc.

Remove the unused code from connections.cc in favour of the new
Task based system as this avoids the race conditions when getting
tasks for connections on a different thread.

Change-Id: If81a83cc23f480988588d077f321d06f5921fa86
Reviewed-on: http://review.couchbase.org/92130
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>