Disable stats connections test under TSAN until its fixed
We know that the code cause TSAN to complain (by design), and
there is ongoing effort to fix that. Disable the test under TSAN
until the code is fixed so that we may let TSAN check that we
don't introduce regressions elsewhere in the code.

MB-29108 tracks the TSAN issue

Change-Id: Idb108174ba5aee7c14a0e6aec3fcf2d6971e4694
Reviewed-on: http://review.couchbase.org/92291
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>