MB-29184: Introduce the KVStoreBench fixture
We introduce the new KVStoreBench fixture. As part of this change
we add the 'Scan' benchmark to measure the performance of a scan
in a simulated DGM scenario (i.e., we bypass the HastTable and we
go to the storage for every document lookup).

Here an example of output when running the benchamark under both
Couchstore and CouchRocks for 100k items:

Run on (8 X 2500 MHz CPU s)
2018-04-10 15:22:13
------------------------------------------------------------------
Benchmark                           Time           CPU Iterations
------------------------------------------------------------------
KVStoreBench/Scan/100000/0   88858675 ns   87575667 ns          9   1115.11k items/s Couchstore
KVStoreBench/Scan/100000/1  256992229 ns  253830333 ns          3   384.73k items/s CouchRocks

Change-Id: Iaccd3b076706819b39e9f6686ec7ac6b04ee3221
Reviewed-on: http://review.couchbase.org/92494
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>